# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "maeum-pet-store",
    "description": "Delicious fastify.js boilerplate",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
...
      "ICategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "category id"
          },
          "name": {
            "type": "string",
            "description": "category name",
            "maxLength": 200,
            "minLength": 2,
            "pattern": "^[a-zA-Z0-9_-]+$"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "description": "CategoryDto of Pet, Store",
        "title": "ICategoryDto"
      },
...
      "IImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "id"
          },
          "refId": {
            "type": "string",
            "description": "reference entity id"
          },
          "metadata": {
            "type": "string",
            "description": "additional metadata for image and entity",
            "maxLength": 2000
          },
          "hash": {
            "type": "string",
            "description": "hash filename to prevent duplicate filenames",
            "maxLength": 512
          },
          "source": {
            "type": "string",
            "description": "the filename at the time of upload",
            "maxLength": 1000,
            "minLength": 2
          }
        },
        "required": [
          "id",
          "refId",
          "hash",
          "source"
        ],
        "description": "CategoryDto of Pet, Store",
        "title": "IImageDto"
      },
...
      "IPetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "id"
          },
          "name": {
            "type": "string",
            "description": "Pet name",
            "maxLength": 200,
            "minLength": 2
          },
          "category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ICategoryDto"
            }
          },
          "tag": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ITagDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "tag"
        ],
        "title": "IPetDto"
      },
      "ITagDto": {
        "title": "ITagDto",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "category id"
          },
          "name": {
            "type": "string",
            "description": "category name",
            "maxLength": 200,
            "minLength": 3
          }
        },
        "required": [
          "id",
          "name"
        ],
        "description": "TagDto of Pet, Store"
      },
...
      "IPostUploadImagePetBodyDto": {
        "type": "object",
        "properties": {
          "additionalMetadata": {
            "type": "string",
            "description": "additional metadata for image and entity",
            "maxLength": 2000
          },
          "file": {
            "$ref": "#/components/schemas/TFileUpload"
          }
        },
        "required": [
          "file"
        ],
        "title": "IPostUploadImagePetBodyDto"
      },
      "TFileUpload": {
        "title": "TFileUpload",
        "type": "string",
        "format": "binary"
      },
...
    }
  },
  "paths": {
...
    "/pet/{id}/uploadImage": {
      "post": {
        "operationId": "upload-pet-image",
        "summary": "uploads an image",
        "tags": [
          "Pet"
        ],
        "description": "upload an image for Pet",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/IPostUploadImagePetBodyDto"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "6f1a29b5-cdc6-4023-a63f-7696d963d09a"
            },
            "in": "query",
            "name": "tid",
            "required": true,
            "description": "request id from client"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IImageDto"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    },
                    "validation": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "data": {},
                          "schemaPath": {
                            "type": "string"
                          },
                          "params": {}
                        },
                        "required": [
                          "message",
                          "schemaPath"
                        ]
                      },
                      "description": "information of validator result"
                    }
                  },
                  "required": [
                    "message",
                    "validation"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/pet/{id}": {
      "delete": {
        "operationId": "delete-pet",
        "summary": "Delete Pet",
        "tags": [
          "Pet"
        ],
        "description": "Delete Pet using id",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "6f1a29b5-cdc6-4023-a63f-7696d963d09a"
            },
            "in": "query",
            "name": "tid",
            "required": true,
            "description": "request id from client"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "category id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPetDto"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    },
                    "validation": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "data": {},
                          "schemaPath": {
                            "type": "string"
                          },
                          "params": {}
                        },
                        "required": [
                          "message",
                          "schemaPath"
                        ]
                      },
                      "description": "information of validator result"
                    }
                  },
                  "required": [
                    "message",
                    "validation"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "read-pet",
        "summary": "Read Pet",
        "tags": [
          "Pet"
        ],
        "description": "Read Pet using id",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "6f1a29b5-cdc6-4023-a63f-7696d963d09a"
            },
            "in": "query",
            "name": "tid",
            "required": true,
            "description": "request id from client"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPetDto"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    },
                    "validation": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "data": {},
                          "schemaPath": {
                            "type": "string"
                          },
                          "params": {}
                        },
                        "required": [
                          "message",
                          "schemaPath"
                        ]
                      },
                      "description": "information of validator result"
                    }
                  },
                  "required": [
                    "message",
                    "validation"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  }
}